Restructure composition of menu popups
