+++ src/components/general/PopupMenu.module.css
@@ -26,16 +26,6 @@
   border-radius: 4px;
 }
 
-.popup:hover .icon {
-  background: var(--secondary);
-  border: 1px solid var(--secondary-dark-2);
-  transform: scale(1.07);
-}
-
-.popup:active .icon {
-  transform: scale(0.95);
-}
-
 .popup:hover path {
   fill: var(--white);
 }
@@ -47,12 +37,6 @@
 
 .popup.disabled {
   cursor: default;
-}
-
-.popup.disabled:hover .icon {
-  background: none;
-  border: 1px solid var(--neutral-light-6);
-  transform: none;
 }
 
 .popup.disabled:hover svg {
@@ -67,12 +51,6 @@
 .menu {
   background-color: var(--white);
   text-align: start;
-}
-
-.icon {
-  border-radius: 4px;
-  border: 1px solid var(--neutral-light-6);
-  overflow: hidden;
 }
 
 .item {
