+++ src/containers/nav/desktop/NavIconPopover.tsx
@@ -1,12 +1,15 @@
-import { IconCrown, IconDashboard, IconSettings } from '@audius/stems'
-import { ReactComponent as IconKebabHorizontal } from 'assets/img/iconKebabHorizontalAlt.svg'
-import IconPopup from 'components/general/IconPopup'
-import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import React from 'react'
+
+import { IconCrown, IconDashboard, IconSettings } from '@audius/stems'
+
+import { ReactComponent as IconKebabHorizontal } from 'assets/img/iconKebabHorizontalAlt.svg'
+import { PopupMenuIconButton } from 'components/general/PopupMenuIconButton'
+import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import { getAccountIsCreator } from 'store/account/selectors'
 import { useSelector } from 'utils/reducer'
 import { AUDIO_PAGE, DASHBOARD_PAGE, SETTINGS_PAGE } from 'utils/route'
 import { removeNullable } from 'utils/typeUtils'
+
 import styles from './NavIconPopover.module.css'
 
 const messages = {
@@ -47,7 +50,7 @@ const NavIconPopover = () => {
 
   return (
     <div className={styles.headerIconWrapper}>
-      <IconPopup
+      <PopupMenuIconButton
         menu={{ items: menuItems }}
         icon={<IconKebabHorizontal />}
         popupClassName={styles.scalingWrapper}
